play game

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,8 +19,8 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('total_points')->nullable();
-            $table->integer('tickets')->nullable();
+            $table->bigInteger('total_points')->nullable();
+            $table->bigInteger('tickets')->nullable();
             $table->rememberToken();
             $table->timestamps();
         });
